Project: only support Git as VCS for new projects (#10114)
* Project: only support Git as VCS for new projects

This is the continuation of a soft deprecation. We already remove mentions to
other VCS different than Git from our documentation.

In this step, we remove the ability to import other repositories other than Git.
However, we are keeping the code around to avoid existing projects to break when
building. In the future, we will remove this support completely.

* Allow projects to update their `repo_type`