Unify feature check for organization/project (#8920)
We are overriding classes and skipping tests con .com just because on .com we require a subscription for some features.
Now that we have subs on .org we can share more code between both sites. 
We have a setting to enable a set of features, this is for users that don't have a subscription (.org), on .com this will {} to rely entirely on the features from the subscription.