Addons API: update required/non-required fields
After using this endpoint for some weeks and creating/designing the addons I
found there are some attributes that are not going to be required anymore and
there are others that we will need to add.

This is the begining of this work that removes the ones we are not using at all
from the Javascript client for now.