Build: allow multi-line commands on `build.commands` (#10334)
* Build: allow multi-line commands on `build.commands`

This is another attempt to fix this issue. It follows the suggestion posted by
@bensze01 in https://github.com/readthedocs/readthedocs.org/issues/10103. It
looks simple and I think it makes sense to give it a try since it doesn't seem
to be harmful.

I tested it locally with the branch
https://github.com/readthedocs/test-builds/tree/build-commands-multiline and it
worked fine. Note that without this PR the branch linked fails.

Related #10133
Related #10119
Related #10206

External PRs that should be solved by this one:
* https://github.com/pypi/warehouse/pull/12953
* https://github.com/pypi/warehouse/pull/12953

Closes #10103
Closes #10208

* Build: use `;` to separate PREFIX from COMMAND