Version: new field `addons`
This `addons` field will be used to tell El Proxito this version was built using
`build.commands`. Then, El Proxito will use this field to inject the HTTP header
that tells Cloudflare that it should inject the new JS client.

This commit is a replacement of
https://github.com/readthedocs/readthedocs.org/pull/10219 to avoid the
time-consuming query that checks the config file of the latest build for the
version.

Closes #10331