Deprecation: remove support for Sphinx 1.x
We are installing Sphinx 1.x based on a feature flag.
There are around 900 projects still using 1.x, based on
https://ethicalads.metabaseapp.com/question/250-projects-using-sphinx-timeserie

Note this feature flag is enabled by default for new projects starting on Oct.
20, 2020

We may need to send them an email about this and informing them how to create a
`requirements.txt` file to pin their dependencies. I don't think that the
migration to a config file v2 would enforce them to pin Sphinx.

Related #10342