Deprecation: remove code for config file v1 and default config file
This is a continuation of the work done for

* https://github.com/readthedocs/readthedocs.org/issues/10351
* https://github.com/readthedocs/readthedocs.org/issues/10342
* https://github.com/readthedocs/readthedocs.org/issues/10348
* https://github.com/readthedocs/readthedocs.org/pull/10355

In that work we added notifications, warning messages and finally a feature flag
to start failing the builds without a configuration file or using v1.

This commit removes all the code for config file v1 and building without a
configuration file. Due to that, it also removes the code that allows users to
use the older Docker images (testing, latest, stable)

This work can be merged once we enable the feature flag to fail the builds for
all the projects and we are happy with the results.