Build: avoid breaking builds when a new argument is added to a task
For https://github.com/readthedocs/readthedocs.org/pull/10289/
we are going to need to pass a new argument (build_api_key).
And since we deploy webs first,
builders will have the old task that doesn't match the new signature,
and the task will fail.

To avoid this, we can just accept any kwargs,
this obviously only works if the change is backwards compatible
with the old code from the builders (in this case it will be).