Fix display issues with project creation config page (#10398)
This addresses some display bugs from commercial, where the nested order
of elements was swapped and there are already conflicts between usage of
`<code>` and `<pre>`.

Instead, this reverts to my suggestion from review, and uses the Pygments
wrapping element `div.highlight` to style code block content. This avoids
conflicts between code/pre use and styling and provides a better looking
border and padding for block content.

This also survives the translation to the commercial styling better and
doesn't affect any other case in the UI where `<code>` is used to style
inline content instead of block content.

There is a required PR on commercial too.