Celery: increase frequency of `delete_closed_external_versions` task
We have 255k versions that should have been deleted already, but we are
executing the task with a limit of 200 version. This makes the task to not catch
up with all the versions that has to delete, and we keep accumulating them.

This commit increases the frequency from "once a day" to "every 3 hours". It
will help to catch up and to keep our database clean of closed PRs.

Related https://github.com/readthedocs/readthedocs-ops/issues/1291