Feature flag: remove unused ones (#10423)
* Feature flag: remove `USE_SPHINX_BUILDERS`

We can safely remove this feature flag since it's marked as `default_true=True`
and `future_default_true=True` in both platforms. This means that it applies to
all the projects.

* Feature flag: remove `SPHINX_PARALLEL`

It was a failed attempt to make Sphinx to run fast.

* Feature flag: remove `CLEAN_AFTER_BUILD`

It became a Django setting and it's integrated in the application now.

* Feature flag: remove `CANCEL_OLD_BUILDS`

It's enabled by default to all the projects and it has been working fine.

* Feature flag: remove `LIST_PACKAGES_INSTALLED_ENV`

This was an attempt to show debugging information. This can be achieved now with
`build.jobs` in case the user want to display this data.

Internally, we are storing some of this data in BuildData from `telemetry`
database.

* Feature flag: remove `VCS_REMOTE_LISTING`

We have been using this by default in all projects.

* Feature flag: remove `USE_RCLONE`

We have been using this for a long time and we are happy.

* Feature flag: remove `CACHED_ENVIRONMENT`

We removed this code completely and we are not doing this anymore.

* Feature flag: remove `ALLOW_DEPRECATED_WEBHOOKS`

It's not used. Only leftovers.

* Feature flag: remove `USE_TESTING_BUILD_IMAGE`

Only used by 1 projects. Its last build was 2 years ago.

* Feature flag: remove `LIMIT_CONCURRENT_BUILDS`

We have been using this by default to everybody for months now and it's working
fine. It's safe to remove.

* Test: replace `storage.sync_directory` with `storage.rclone_sync_directory`

* Remove leftovers from deprecated webhooks

* Test: check for the original builder

* Test: fix them all (hopefully)

* Lint

* More linting

* Test: mock required function