Proxito: easy migration to custom path prefixes
With this we would be able to start generating
links using the path-prefix while serving docs with the old
proxito implementation.

And this also allows us to start using the new implementation
with just enabling the feature flag for a project.
This is, without having to unset the `urlconf` attribute.
This help us to have zero downtime while migrating projects.