Addons: handle resolver API's exceptions
If there is an exception on a valid project, that means the project exists but
maybe the translation, version or the filename does not. In this case, we return
all the data available but nulling (`null`) values that require a valid version
to work.

I made this decision because even on 404 pages we will want to show the flyout,
for example, so we still need most of the data to be returned.