Update gold docs to reflect cross-site cookie reality
Remove from our landing pages that being a gold member
or supporter makes RTD ad free. We aren't removing that
capability, but the cross site cookies needed to make that
work do not work on Safari or Firefox.
We'll leave the functionality in for Chrome users,
but we aren't going to sell it since it doesn't work widely.

Basically, the API requests from *.readthedocs.io or
user custom domains cannot make an authenticated request
to readthedocs.org.