Addons: improve db query when adding HTTP header from El Proxito (#10461)
We don't need the `Project` instance. We were using to check for a Feature flag,
but now we are only adding the header when the project is using
`build.commands`.

If we ever require the `Project` instance again, we can do a `.select_related`
from the `Version` object. That way, we still keep doing 1 query only.

Closes #10458
Related #10456