Feature flag: remove `DONT_CREATE_INDEX`
Remove the usage of `DONT_CREATE_INDEX` feature flag that only 1 project is
currently using.

Besides, this commit removes the auto-creation of the `index.html` file when
building with MkDocs or Sphinx. We will be checking at "Uploading" status if
there is no `index.html` file and we will fail the build in that case.