Addons: handle API exceptions from unresolver (#10427)
* Addons: handle API exceptions from unresolver

The new unresolver raises exceptions for different reasons. For example, on 404
pages. We need to prepare the API to handle these cases and decide what to do on
each of them.

Closes #10399

* Addons: handle resolver API's exceptions

If there is an exception on a valid project, that means the project exists but
maybe the translation, version or the filename does not. In this case, we return
all the data available but nulling (`null`) values that require a valid version
to work.

I made this decision because even on 404 pages we will want to show the flyout,
for example, so we still need most of the data to be returned.

* Test: make API test to work again, remove a field

* Addons: improve "active and built Versions" query (#10455)

This matches the query executed on FooterAPI.
It will also helps with the performance of the query.

Closes #10429