DB: clean up old data using raw SQL from Django
We are facing an issue with this query because it takes too long to
execute (more than 30s) making our DB to kill the query. This is because Django
performs a `SELECT` first to be able to trigger pre_ and post_ delete signals on
each object delete.

We don't really need this here, so we are using raw SQL to bypass this and make
the query to execute faster. This is not ideal, but we didn't find a better approach.