Version warning banner: disable it for project not using it already
We disccussed about this in https://github.com/readthedocs/meta/discussions/125

Basically, I want to disable it because it's more confusing than useful
currently. Besides, it will be completely replaced by the new addons work we are
doing.

This is the first step we are taking here: stop keep confusing users. The
following step would be to contact (if required) projects already using this feature and
migrate them to the new addons.

This commit includes a migration that creates the feature flag and assign all
the projects that are already using `Show version banner` to that flag.