Builds: set scale-in protection before/after each build
We use AWS scale-in protection to protect our instances from being scaled-in
automatically while running a build. This avoids killing the build in the middle
while it's running.

Besides, it could help us to be more aggressive to scale-in if we consider
without degrading user experience.

Related https://github.com/readthedocs/readthedocs-ops/issues/1324