OAuth: catch `TokenExpiredError` exception
There is nothing we can do in this case. So, we are logging the exception
instead of sending it to Sentry as an error.

In the future we could handle this exception differently and communicate these
errors to the users so they can re-install Read the Docs app in their GH accounts.

Sentry: https://read-the-docs.sentry.io/issues/3526635712/