Build: remove PyPy support (#10514)
We introduced PyPy support as an alternative way to make Sphinx builds faster in
https://github.com/readthedocs/readthedocs.org/issues/3870

We enabled it in some projects and we tried to communicate this to some CPU
intense projects. However, nobody kept using it after doing some experiments
with it.

I checked our logs and Telemetry data and I found 0 projects using `pypy` in
their builds. So, I'm happy to remove its support and simplify our product as
its maintainance as well.

Even the PyPy project is not using it when building their own documentation:
https://readthedocs.org/projects/pypy/builds/21201174/