Docs: swap around content for configuration files (#10517)
* Alter search ranking of configuration file content for now

This alters the search ranking of two pages: the configuration file
index page and the guide on adding a configuration file. The guide
covers all of the configuraiton file basics, where the current index
page is more of a "our vision on configuration files". Users should see
the guide content first, I've noticed it's hard to surface in search
though.

Long term, I think the guide content should replace the configuration
file index page, which should be in turn moved somewhere else. I tried
this here, but I got a little lost on what page structure we're aiming
for.

* Move the configuration file pages and promote the tutorial more

* Add config file tutorial to tutorial list index as well