Build: use `only-if-needed` pip's strategy when installing package
We are currently using `--upgrade-strategy eager` which *upgrades all the
packages** installed even if they were previously installed to a particular
pinned/defined version by the user.

Using `only-if-needed` will upgrade only the dependencies that are defined in
the package and higher version are required.

This new behavior will avoid unexpected results/upgrades.

Closes #9752