Build: skip duplicated commands
Sometimes it happens the web servers are congested and the builder retries the
API call to save the command in the database. However, one of the first attempts
finally ended up working resulting in the command being saved twice.

This small chunk of code checks for the existence of the command before saving
it into the database and logs a warning and return 204 if it already exists,
instead of re-saving it again.

It's not a perfect solution, but it could help under similar circumstances.

Related https://github.com/readthedocs/readthedocs.org/issues/10567