Proxito: test new implementation more broadly
With this we can now exclude some projects from using the
new implementation while testing it more broadly
by changing the date of the feature.

When we are ready to enable it for all projects, we can
remove the USE_UNRESOLVER_WITH_PROXITO, and leave the other
setting while we migrate the missing projects.