Requirements: revert upgrade to `psycopg==3.x`
This package works fine and the upgrade was transparent. However, it seems that
New Relic agent is not prepared to collect data from v3.x and send it to their
servers.

https://docs.newrelic.com/docs/apm/agents/python-agent/getting-started/instrumented-python-packages/#instance-details

> The Python agent should be able to track database queries for any Python
> DB-API 2.0 compliant modules

`psycopg3` is compatible with DB-API 2.0, but for some reason doesn't work.
I'm reverting it for now and we can take a look later.

Note this commit also upgrades Django, due to a security issue that was
reported:

https://docs.djangoproject.com/en/4.2/releases/4.2.5/

Reverts #10667