Addons: allow users to opt-in into the beta addons
I created the URL, models, view and form to prepare ourselves for the beta
addons. Currently, if the `Project.addons` fields is not null we add a HTTP
header to tell Cloudflare to enforce the new beta addons on this project.

In the future, we can use each of the `ProjectAddonsConfig` field to let the
user decide which of the addons they want to enable/disable and change their
configurations.

* Related: https://github.com/readthedocs/ext-theme/issues/212