Addons: get final project (e.g. `subproject`)
Use `unresolved_url` to get the final project requested from the URL. This is
needed for subproject since it's the only place where it's present.

Previously, we were getting the project from the `unresolved_domain` which does
not give us the `subproject` because it's not present in the domain itself.

Tests will be added in
https://github.com/readthedocs/readthedocs.org/issues/10685

Closes https://github.com/readthedocs/readthedocs-corporate/issues/1622