Custom domain doc improvements (#10719)
* Drop custom domain mermaid graphic

I don't feel this improves the user understanding of how custom domains
work, and seems to be a pain to update to make it more helpful. Removing
this for now.

* Explain our default subdomains better

Mention organization name as well, but don't surface the word "slug" as
"name" is more relateable to the user.

* Remove custom domain guidance for open source projects

This stands out as misplaced content for our feature page on custom domains.

* Rework set up notes and "how custom domains work" into features section

This drops "how custom domains work", which was some duplicate content
of our how to set up custom domains guide. Content that leaned more
towards describing the features of custom domains was reworked into a
feature section that lists the features that make up custom domain
support.

* Backwards link syntax

* Don't use page title for CDN doc

* Remove missed extra word

Co-authored-by: Santos Gallegos <stsewd@proton.me>

* Update docs/user/custom-domains.rst

Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>

---------

Co-authored-by: Santos Gallegos <stsewd@proton.me>
Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>