Remove addons response for last build
We noticed a big spike in time spent in build table queries today and
traced the query back to the addons response. The addons data structure
includes a list of the latest builds for each version.

I've disabled that for now, but it's not yet clear why this became an
issue on Fri, or why the cache doesn't seem to be helping much here.