Build: don't pre-install pip and setuptools in images
We hit a problem with Python 3.12 and the outdated version of pip
that we have pre-installed.

Python ships with a vendored version of pip,
and we install the latest version of pip and setuptools on each build
before installing the requirements.

The only breaking change here would be for projects using
build.commands, setuptools won't be available anymore,
pip is vendored by Python, so that shouldn't be a problem.
I think we are fine not installing setuptools,
build.commands is in beta, so we can get away with this :)

Fixes https://github.com/readthedocs/readthedocs.org/issues/10832