DB: create an index for `builds_build` table to improve Addons API
After doing some research on #10831, I think that adding an index to this table
will help a lot to improve queries performance on Addons API.

This PR also fixes the `version.builds.last()` query to get the `Build` object
that is finished and successfull. Otherwise, we will be returning a build that
does not match with the HTML on S3.

I think we should close #10831 and revert the commit on `rel*` branches.