Addons: pre-fetch the custom `Domain` instance to resolve URLs
The `resolver.resolve` method that resolves the URL **always** check for the
canonical `Domain` of the canonical `Project`. This means that resolving
the URLs for multiple `Version` objects results in querying the exact query over
and over again.

Instead of repeating this query, we are passing the `Domain` object to
`resolver.resolve` method to avoid hiting the DB again since we already known
what's the canonical domain for that project, in case it has one.

I found this issue while tracking down why there were around ~150 queries to the
`projects_domain` table when hitting `/_/addons/` API endpoint on New Relic.