Addons: normalize search filters
Currently, we have two filters on shown by default in the search modal.
However, they are not super useful to work together.

I changed this behavior (for now) to be:

* Default: search on the current project
* Include subprojects: also include results from subprojects

Note this will be configurable by the user once we expose the addons admin page
in https://github.com/readthedocs/ext-theme/issues/211

Closes https://github.com/readthedocs/addons/issues/209