Use a custom `models.Manager` to implement de-duplication
Use `Notification.objects.add()` to create a notification with de-duplication
capabilities. If the notification already exists, it updates all the fields with
the new ones and set the `modified` datetime to now.

This commit also migrates the "Your primary email address is not verified" to
the new notification system, which it was when I implemented the de-duplication logic.