Notifications: small fixes found after reviewer (#10996)
* Notifications: small fixes found after reviewer

Use single `{}` since we are using `.format()` to format these strings.

* Handle missing `.format()` key

Avoid internal error and log the exception so we can figure it out how to solve
it. This happens when the `Notification` does not have _all_ the required `format_values`.

* Protection agasint XSS when rendering notifications

See https://github.com/readthedocs/readthedocs.org/pull/10922#discussion_r1440756414

* Test for missing key in format values and XSS protection

* Update common/

* Lint

* Document we only support `str` and `int` for now in `format_values`

We don't support nested dictionaries in `format_values` or random objects.
Only `str` and `int`. That should be enough for now.

Skip all the values that are not `str` or `int` from the format values to render
the messages.

* Typo