Resolver: don't use str.format
We were using .format over a user controlled string, the project prefix.
This can be used to traverse to some private variables in some cases
(this isn't the case here), or to cause a DoS by using an expensive
format operation (`{language:>999999999999}`).

This isn't a security issue, since we don't expose the prefix to users yet.