PR previews: match repository privacy level
I was just adding a warning at the start,
but then I realized that we can just take the privacy level
of the repository into consideration as well.

If the project was manually imported,
we have no way of knowing the privacy level of the repository
(well, maybe we could infer it from the repo URL and if all versions are
public).

Ref https://github.com/readthedocs/readthedocs.org/security/advisories/GHSA-pw32-ffxw-68rh