Embed API: fix regex patterns for allowed external domains
The patterns were not ending with `$`,
so any domain that started with the allowed domain would be allowed
(docs.python.org.example.com).

This isn't a security issue, since including content from a domain
that isn't allowed is not differently than including content from
a domain that is hosted on RTD (users shouldn't allow including content
from projects they don't trust).
This is mostly to prevent abuse.