Initial example of using notification system for view-specific errors
I found myself rebuilding these messages in template code and figured it
would be a good excuse to further consolidate the patterns we're using.

In short, these messages are emitted at view time and are not saved to
the database. The view time logic is similar as notifications however.

A few questions stand for me:

- I like centering all of our notifications/errors around exceptions, as
  it retains an opportunity to emit exceptions from wherever we might
  want to move some of this logic. Arguably, this logic should not exist
  on a single view and could be moved on to modeling/queryset classes.
- I don't have a strong opinion on whether it's important for these
  messages to exist in the registry. I see arguments for both, but I
  chose to use the registry so that we can consolidate logic there --
  format values and message lookup, etc.

I'm looking for input on these patterns.

On the template side, the display is now just a simple `if` and template
logic similar to the current notification display, using the header/body
render methods.