Refactor documentation navigation
This is an initial proof of concept of a concept-based TOC,
instead of using Diataxis.

This shows a lot more where we're missing various pieces of content,
and overall I think really improves our documentation structure.

A big benefit downstream of this is making our page titles more explicit,
so users know what each page has on it.
I've started doing this very lightly so you can get a sense,
but it makes things way more clear.