Dev: Default RTD_DJANGO_DEBUG to False if not set
With the docker compose settings file, it is currently necessary to
explicitly set RTD_DJANGO_DEBUG to a false-y value to clear the DEBUG
flag.

Change the default value to False to make the setting more intuitive.

The docker invoke script in common passes RTD_DJANGO_DEBUG=t by default
so the effective default behavior is unchanged.

Signed-off-by: Peter Hoyes <Peter.Hoyes@arm.com>