Notification: fix `choices` rendering for `INVALID_CHOICE`
While taking a look at #11189 I found that the rendering was broken.
I found the `with` + `yield` pattern pretty complex (I have a note about this in
the code) and it seems that complexity is confusing enough to make this hard to
dealt with.

Anyways, I fixed the problem for now and I added a test case. We can come back
to refactoring this in the future with more time.