Allow setting Allauth provider secrets from host system (#11194)
* Allow setting Allauth provider secrets from host system

This requires

- https://github.com/readthedocs/common/pull/209

It is modeled after the Stripe setting pattern. There is a stronger pattern here at some point, this is a bare minimum for now to unblock us from not having these settings on our local environments.

You can use `direnv` to automatically export (and encrypt if you are so inclined) these secrets for your local envs. The env vars pass through Docker with common/#209 above.

* Add some docs too

* Note direnv too

* Drop unused key env vars/settings for allauth pass through

* Update common

* Drop bitbucket provider settings/env vars

* Update common