APIv3: add more generic fields
Add `aliases` field to the `VersionSerializer` that's used to serialize
`Version` objects. This field is a list of versions the current version "points
to". Example `latest` version "points to" `origin/main`, and `stable` version
"points to" `v3.12`.

This is known in the code as "original stable/latest version".

Besides, this PR adds `versions.stable` and `versions.latest` fields to the
addons API which will be useful to solve front-end issues in a more generic way.