Expose generic fields (`projects.translations`, `versions.active`)
This is the pattern I want to have here, serializing the full object instead of
a few small set of fields. This is a lot more generic, simplifies the
interaction between the API and JS Addons.

Besides, it exposes the same objects as the regular APIv3 which won't change
over time.