Add error view for error handling and error view testing (#11263)
* Add error view for error handling and error view testing

I was having trouble testing error views and templates locally and found
at some point we added a `/500` for testing the 500 error page. I
expanded on this to create a view that could emit any error status code
and relevant template. Afterwards, it seemed like we had a few redundant
views doing similar things, so I expanded the dashboard error handlers
to all use the same view as well.

Changes here are:

- When using the new dashboard, dashboard error templates are in the
  `/errors/dashboard/` template path. There will be a duplicate set of
  templates for Proxito errors and documentation errors, in
  `/errors/proxito/` eventually.
- This swaps out the URL `handler500`, and defines the missing handlers
  in the URL module. Mostly, this loads these error templates from a
  more structured path instead of top level `/404.html` templates.
- Adds a debug view/URL for testing arbitrary error views/templates in
  debug mode.

This does not touch proxito error pages yet.

* More linting

* Support all request methods in ErrorView

* Add support for spam project response/410

This is just the dashboard response view for spam projects.

* Symlink 403 to 401 template

The 401 template has copy closer to 403 already, and I don't believe we
are emitting a templated 403 response anywhere either.