Adapt `__str__` and `__repr__` methods for DB
This commit removes some debugging functionability in favor of production DB.
`__str__` and `__repr__` methods won't be so descriptive now since we are
removing some information from their rendering. This is because to render them
properly we need to hit the DB multiple times in the worst case --generating 500
on some user requests that need to be logged in Sentry/New Relic.

There are better ways for this: disabling logging on production and enabling it
on DEBUG + Django Shell, but that requires more extra work that doesn't seems
super priority right now. We can come back later and add them as we need them if
we want.

Closes #10954